feat: use masl directly for token exchange
