--- js/packages/teams-dialogs/src/credential/onBehalfOfCredentialAuthConfig.ts
@@ -1,40 +1,0 @@
-/**
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export type OnBehalfOfCredentialAuthConfig = {
-  /**
-   * Hostname of AAD authority.
-   */
-  authorityHost: string;
-
-  /**
-   * The client (application) ID of an App Registration in the tenant
-   */
-  clientId: string;
-
-  /**
-   * AAD tenant id
-   *
-   * @readonly
-   */
-  tenantId: string;
-} & (
-      | {
-          /**
-           * Secret string that the application uses when requesting a token. Only used in confidential client applications. Can be created in the Azure app registration portal.
-           */
-          clientSecret: string;
-          certificateContent?: never;
-      }
-      | {
-          clientSecret?: never;
-          /**
-           * The content of a PEM-encoded public/private key certificate.
-           *
-           * @readonly
-           */
-          certificateContent: string;
-      }
-  );
